перенес загрузку и отправку данных в отдельные модули
